Unstable

+++ alframework/ml_interfaces/hippynn_interface.py
@@ -385,15 +385,16 @@ def train_HIPNN_model(model_dir,
                     del arrays[key]
             
             # Remove High energies and forces. 
-            database.remove_high_property(force_key, True, species_key=species_key, cut=remove_high_forces_cut, std_factor=remove_high_forces_std)
+            if force_key != None:
+                database.remove_high_property(force_key, True, species_key=species_key, cut=remove_high_forces_cut, std_factor=remove_high_forces_std)
             database.remove_high_property(energy_key, False, species_key=species_key, cut=remove_high_energy_cut, std_factor=remove_high_energy_std)
             
             print("Array Shapes After Cleaning")
             prettyprint_arrays(database.arr_dict)
-            print(np.max(database.arr_dict['forces']))
-            print(np.min(database.arr_dict['forces']))
-            print(np.max(database.arr_dict['energyperatom']))
-            print(np.min(database.arr_dict['energyperatom']))
+            print(np.max(database.arr_dict[force_key]))
+            print(np.min(database.arr_dict[force_key]))
+            print(np.max(database.arr_dict[energy_key+"peratom"]))
+            print(np.min(database.arr_dict[energy_key+"peratom"]))
                         
             database.make_random_split("valid",valid_size)
             if not(h5_test_dir == None):
